build boxes on qemu/libvirt

+++ packer/scripts/cleanup.sh
@@ -48,10 +48,20 @@ rm -rf /usr/src/vboxguest*
 rm -rf /usr/share/man/??
 rm -rf /usr/share/man/??_*
 
-# Zero out the free space to save space in the final image:
-echo "Zeroing device to make space..."
-dd if=/dev/zero of=/EMPTY bs=1M
-rm -f /EMPTY
+# Discard flag is enabled, but we did not remount/reboot
+# Zero out the free space to save space in the final image,
+if [ $PACKER_BUILDER_TYPE != 'qemu' ]; then
+	echo "Zeroing device to make space..."
+	dd if=/dev/zero of=/EMPTY bs=1M
+	rm -f /EMPTY
+else
+	echo "Full disk write would expand the QCOW2 image - skipping"
+fi
+
+# Run fstrim 'just to be sure' trim works
+echo "trim + 5 second wait"
+fstrim /
+sleep 5
 
 # Remove history file
 unset HISTFILE
